esl_msa.c: Updated 'Purpose' comments of esl_msa_SequenceSubset() to reflect changes made in last revision (398).
Comments below were mistakenly omitted from log of *previous* commit
for revision 398:

Makefile.in: 
Added new 'ssu' argument for 'make' for building easel as a library
within a SSU-ALIGN distribution. 'make ssu' results in miniapps being
copied to renamed executables, for example 'esl-sfetch' becomes
'ssu-esl-sfetch'. The idea is 'ssu-' prefixed files would then
installed in a 'libexecdir' following GNU coding standards. They will
be accessed by SSU-ALIGN scripts but are not meant to be called
directly by users, and we don't want them causing confusion in user's
paths with easel miniapps installed by infernal or HMMER.

miniapps/esl-alimanip.c:
Several minor changes. With --mask-rf and --mask-all the mask file is
now checked to make sure it is of valid length and only '0/1'
characters.

miniapps/esl-ssdraw.c:
Major change regarding the postscript template file: it can now
contain >1 SS diagrams. esl-ssdraw will now check each of the diagrams
for a 'match' to the model used to build the input alignment. A
'match' is defined as a SS diagram with the same consensus length as
the input alignment (non-gap RF length). This is a fragile method.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.0@399 db3e70e4-12e1-0310-b23f-81a6e3d09ba7