Added esl-alimask miniapp, updated miniapps/Makefile.in so it gets built and added esl-alimask.itest.pl test script and esl-alimask.man manual page.
esl-alimask removes columns from a multiple sequence alignment. It can
be run in 5 different modes which are selected based on command-line
usage: 

  esl-alimask [options] <msafile> <maskfile>   (use mask from maskfile)
  esl-alimask [options] -t <msafile> <coords>  (truncate alignment to coords)
  esl-alimask [options] -g <msafile>           (use gap frequencies in aln)
  esl-alimask [options] -p <msafile>           (use post probs (PP) in aln)
  esl-alimask [options] --rf-is-mask <msafile> (use #=GC RF in aln as mask)

Only one usage listed above can be used per execution with the exception of
-p and -g, which can be used in combination with each other.

The code is more complex than I initially intended but I decided all 5
modes are worthwhile and that separate 'usages' was the simplest way
to allow them all within a single miniapp. The goal is to have
esl-alimask be the only miniapp necessary for the task of removal of
columns from an alignment.

With --savemem, if the input alignment file is in Stockholm Pfam
format, the program will run in a memory efficient mode in which the
memory usage is independent of the alignment size, allowing masking of
very large alignments without the need for a lot of RAM.

Updated esl_msa::esl_msa_ReadNonSeqInfoPfam() so that arrays with
counts of number of each residue, and each PP value in each column of
the alignment can optionally be returned. This makes memory efficient
gap-based and PP-based masking possible.

Most of the functionality of esl-alimask is already implemented in
esl-alimanip (without the memory efficient option), but it is more
sloppily written, less efficient, and overly complex. (Two examples:
some types of masking require two successive runs of esl-alimanip, and
esl-alimanip works by rewriting the #=GC RF annotation and then
masking based on that annotation, which can be undesirable in some
situations).  I will soon deprecate the the functionality in
esl-alimanip that is now handled more cleanly and efficiently by
esl-alimask.

Further testing of esl-alimask was done as described in:
Xref: /groups/eddy/home/nawrockie/notebook/9_1217_esl_alimask/00LOG.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@479 db3e70e4-12e1-0310-b23f-81a6e3d09ba7