Added support for NCBI protein database support Version 4.
Broke up esl_sqio into to files, esl_sqio and esl_sqio_ascii.
The file esl_sqio is the main calling interface and will make
calls to the format specific routines.  The file esl_sqio_ncbi
was added for the reading and parsing of blast databases.

When open is called for file XXXX and no format is specified,
easel will first try to open a blast db by looking for XXXX.pin,
XXXX.phr and XXXX.psq files.  If any or all of these files do
not exist, then the file XXXX will try to be opened as an
ascii file ie FASTA, EMBL, etc.  The search order remains the
same with the local directory searched first, then in the order
of the environment variable ex BLASTDB.

To add support for a new file format, a new file esl_sqio_YYYY
would be created.  The a call to the open routine for YYYY would 
be added to esl_sqfile_Open.  If opening of YYYY is successful,
the callbacks to YYYY routines are registered with <sqfile> so
subsequent call to ex esl_sqio_Read or esl_sqio_Position would
then call the correct routines for YYYY.

Some interfaces needed to be changed for example:

esl_sqio_Write(FILE *fp, ESL_SQ *s, int format, int update)

This intefaced added the paramater <update> to specify if the
offsets in <s> needed to reflect the writing of the sequence.
This is used in some unit tests.

Also this interface was added:

const char *
esl_sqfile_GetErrorBuf(const ESL_SQFILE *sqfp)

The error message are formatted by the file type.  For example, line
number is added to several of the error messages when parsing ascii
based files, ie FASTA.  These line numbers make no sense when an
error is encounted for a binary format, so those error messages
contain hex offsets and sequence index numbers.  When an error is
encountered, the calling program can get the error string and print
it with a "%s" format.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@480 db3e70e4-12e1-0310-b23f-81a6e3d09ba7