esl-reformat:
Added --small option to esl-reformat which only works in combination
with '--informat pfam' and output format of pfam or afa. This was
added was so that SSU-ALIGN can convert its (potentially multi-Gb)
Pfam alignments into aligned FASTA on machines without requiring RAM
equivalent to the alignment size.

Additionally, esl-reformat now exits with an error message if the
input format is Stockholm and the output format is A2M, PSIBLAST or
AFA and the input alignment file has > 1 alignments in it. Exit occurs
after the first reformatted alignment is output.

esl_msa:

Small bug fix in how esl_msa_RegurgitatePfam() outputs GS annotation.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@514 db3e70e4-12e1-0310-b23f-81a6e3d09ba7