Fixed several build issues associated with biulding and running hmmer on the AIX operating system using IBM's XLC compiler.
The miniapps changes was the addition of "#include "esl_config.h".  Without
this line the compiler complained about redeclaration of lseek64.

In Makefilethe -f flag to the rm command when cleaning up.  This avoids 
messages when a file is not present.

In configure.ac and aclocal.m4 changes were made in which compiler flags 
were tested and used when compiling with the XLC compiler.  Also added a 
test to see if Altivec/VMX code can compiler on the target machine.  To 
get Altivec/VMX code compiled with some XLC compilers, the cpu architecture 
needs to be set.  This can be done with the configuration flag 
--with-xlc-arc=XXX.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@520 db3e70e4-12e1-0310-b23f-81a6e3d09ba7