esl-alimerge:
Fixed bug that occurs with --list and >2 alignments to merge for which
some but not all alignments have #=GS DE, #=GS AC, #=GR PP, #=GR SA,
or #=GR SS annotation. These are parsed (stored in dedicated arrays)
by esl_msa, and I wasn't properly initializing these to NULL for msa's
in which they are absent.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@594 db3e70e4-12e1-0310-b23f-81a6e3d09ba7