esl-ssdraw:
New features for individual sequence drawings. Nucleotides that are
not equal to the majority-rule consensus nt at each position are now
outlined (boxed). If the consensus nt occurs with frequency > 0.75,
the outline is bold. Counts of outlined nts and also bps that differ
from consensus are included in the legend. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@595 db3e70e4-12e1-0310-b23f-81a6e3d09ba7