easel can now understand "hmmpgmd" format, Farrar's modified FASTA format for the hmmpgmd HMMER server.
It doesn't autorecognize it: you have to specify --informat hmmpgmd.
It also can't write it: that'd take a bit too much work right now, since
it has a special syntax for merging multiple databases into one quasi-FASTA file.

but the miniapps will work on it, including what Rob asked for, which
is sfetching ability.

To index an hmmpgmd file:
   esl-sfetch --informat hmmpgmd --index hmmpgmd.fa

The --index creates an .ssi file. The .ssi file records the file
format as HMMPGMD, so no more --informat command line options are
needed.

So, to fetch: (in hmmpgmd files, sequences are numbers):
    esl-sfetch hmmpgmd.fa 42



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@639 db3e70e4-12e1-0310-b23f-81a6e3d09ba7