Extended esl_stack to include idiom for a read/write command stack shared between threads. One or more threads (clients) can write jobs to the stack; one or more threads (workers) can take jobs off the stack. This meant including appropriate mutexing and conditional waits in the ESL_STACK and its functions; all of which functionality is conditionally compiled when HAVE_PTHREAD is defined.
This idiom is intended to replace the problematic command queue in the
HMMER hmmpgmd daemon.

Added unit test utest_interthread_comm() for the idiom, which also
serves as example code.

Added esl_stack_DiscardedSelected(), for sudden and arbitrary removal
of elements from a stack. This will replace hmmpgmd's function for
clearing all of one client's jobs out of the stack. Added unit test
utest_DiscardSelected() to test it.

Changed HAVE_PTHREADS to HAVE_PTHREAD; this is more consistent with
standard autoconf usage (because the header file is pthread.h).




git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@651 db3e70e4-12e1-0310-b23f-81a6e3d09ba7