Added "#include "esl_config.h" to top of the 7 miniapps/*.c files that didn't already include it.
This fixed the bug referred to in HMMER's trunk's r3496 log by
Sean. That bug caused a unit test failure of esl-alimanip and was a
result of different sizes for the ESL_SQ object in esl-alimanip and in
the easel library (libeasel.a). This was revealed by printing the
'sizeof(ESL_SQ)' in esl-alimanip and in esl_sq.c. I think this was due
to >= 1 type size (off_t?) being defined differently whether or not
esl_config.h was included or not.

While the addition of "#include esl_config.h" to esl-alimanip.c was
sufficient to fix the bug, I added it to all .c files that lacked it,
assuming that it was an oversight and to avoid similar bugs (which
can be hard to track down if you haven't seen this kind before) in the
future.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@668 db3e70e4-12e1-0310-b23f-81a6e3d09ba7