I commit some simple functions that I have used to investigate the calculation of a gapped lambda for different substitution score matrices and gap open and extension penalties. In particular, functions to plot f(lambda) for a sensible range of lambda values. This plots are particularly useful to understand why multiple roots (f(lambda)=0) or no root at all (see Yu&Altschul, 2005) exist for some gap open and extension ranges.
Multiple roots are not a problem because I think I can now prove that only one is sensible (e.g. does not result in negative background or target frequencies). Having no root is, however, a big problem. As expected, roots tend to disappear the more gappy the score system is and the lower relative entropy the substitution matrix has. That is, the more we deviate from the ungapped theory. Things seem to improve a bit when solving the root with the background frequencies constrained to be as the original ones, which further decreases lambda and pushes f(lambda) to be less negative. A prototype of a function esl_sco_gap_ProbifyGivenBG is included in this commit.

git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/castellanos_3.0rc1@671 db3e70e4-12e1-0310-b23f-81a6e3d09ba7