esl_scorematrix module tidied up.   esl_scorematrix_Load() => esl_scorematrix_Set()   esl_scorematrix_SetBLOSUM62 deleted; use esl_scorematrix_Set()   esl_sco_*() => esl_scorematrix_*()
More builtin PAM/BLOSUM matrices added, including PAM240.

Added the esl_scorematrix_experiment driver, a demonstration of how
the Yu/Altschul (2003,2005) procedure is flawed. Standard matrices are
"low precision" matrices, with small integer values and lambdas in the
range of 1/2 to 1/5 bits. Even when there is a known, valid
probabilistic basis for the score matrix, integer roundoff to a low
precision matrix can create a score matrix that cannot be solved using
the Yu/Altschul procedure. Simply increasing the precision (the scale
-s <x> in the experiment driver) to more integer digits is sufficient
to work around the problem. Undoubtedly it is not a coincidence that
Yu and Altschul (2003,2005) scale all their matrices up by a factor of
500. This is fine so long as you know the probabilistic basis of your
matrix and can recalculate it at increased precision -- but of course
the whole point of Yu/Altschul is to *deduce* the probabilistic basis
from the score matrix. 

Yu and Altschul appear to be saying that they know that their
procedure can find invalid solutions, but that this isn't a problem
because it suffices to keep searching in lambda until a valid solution
is found. This isn't correct. In the examples I'll soon write up in
notes and esl_scorematrix documentation, their procedure finds the
correct lambda, but the p_ij, f_i, and f_b's nonetheless include
negative values, and there's no other solution to find (a matrix
inverse is unique).  Their proof shows that there is not *more* than
one valid solution, which doesn't mean that one exists; it appears
they've jumped the gun on assuming that a solution exists. 

I should check this with Altschul now that I have a small example of
the problem.

This means the procedures I'm using in phmmer/jackhmmer are
unreliable, and must be changed.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@681 db3e70e4-12e1-0310-b23f-81a6e3d09ba7