Upgraded Stockholm parser now in draft, now compilable (esl_msafile_stockholm.c).
API changes on:
  esl_msa_AddComment()
  esl_msa_AddGF()
  esl_msa_AddGS()
... each now takes <p>,<n> style args, for either unterminated buffers
or NUL-terminated strings. To convert any call without thinking about it,
just use -1 as the <n> argument.

Added:
  esl_memtod()
  esl_memtof()
  esl_mem_IsReal()
... memory buffer variants of strtod(), esl_str_IsReal().

Added:
  esl_msafile_stockholm_GuessAlphabet()
This is a stripped-down parser that only counts residues, and doesn't
try to make an ESL_MSA. We'll switch over to using this approach to
guess the alphabet for other MSA formats too.  The issue we're fixing
this way: error reporting. In the previous approach (read MSA in text
mode and cache it; guess alphabet; convert ESL_MSAFILE to digital
mode; then upon first read, instead of parsing, retrieve already-read
MSA from its cache, and digitize it) if the MSA contained an invalid
residue, we didn't notice until we tried to recover it from the cache
and digitize it. By then we have no idea where in the input stream it
occurred (linenumber). In the new regime of MSA parsing, we're 
guaranteeing that the ESL_MSAFILE returns in a very particular state
upon eslEFORMAT errors, to enable strong diagnostic output: including
having afp->linenumber be the line on which the parse error occurred.





git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@687 db3e70e4-12e1-0310-b23f-81a6e3d09ba7