Added "--replace <s>" option to esl-reformat, where <s> must match "<s1>:<s2>" with equal number of characters in <s1> and <s2>. Any character in <s1> in the input sequence file is replaced with its counterpart at the identical position in <s2>.
esl_msa.c and esl_sq.c: corrected documentation (in Purpose field) for
a few functions that listed 'gap' characters as '-_.', omitting '~'
when in fact the code considers '~' a gap in these functions. There
may be more cases of this because for text mode MSAs gap characters
are often defined within each function or set as a function parameter
by the caller. Is it worth #define'ing the 'gapchars' (and probably
also a gap-plus-missing-chars' string in easel.h?  


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@689 db3e70e4-12e1-0310-b23f-81a6e3d09ba7