Fixed bug that caused large SSI files to be written incorrectly. Specifically, the sequence length <len> and disk offset <d_off> values were invalid for ESL_PKEY structures storing information on individual sequences. This only occured when 'external' sort mode had been triggered, which only happens if the size of the SSI index exceeds 256 Mb.
The bug occured in the parse_pkey() helper function that is only used
for writing SSI files in external sort mode. This function parses a
<buf> string from an external primary-key tmpfile, setting the <key>,
<r_off> and <len> values of a ESL_PKEY object. The bug was that it
neglected to parse and set the <d_off> value and consequently
incorrectly set <len> as the <d_off> value parsed from <buf>. This
meant that when a SSI index written in this way was subsequently
read, the <L> value for each sequence was being read and set as the
what should have been the <d_off> value. 

I found the bug when running Infernal's dcmsearch against an
SSI-indexed RFAMSEQ 10.0 database, which is about 170 Gb. The SSI file
only barely exceeds the minimum size for external sorting, which
explains why this bug was not found and fixed long ago.  Infernal's
dcmsearch, which determines the total sequence length of a file by
summing <L> for each individual sequence was determining RFAMSEQ
10.0's total length to be about 5.4 exabytes, which (thankfully) is
incorrect.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@693 db3e70e4-12e1-0310-b23f-81a6e3d09ba7