esl-alistat converted to new msafile interface.
However, I wasn't able to convert the small memory Pfam reading
functions -- at least, I don't see how to do it in a reasonable amount
of time, and I'm already peeved at how long the upgrade is taking.  So
instead, I'm going to leave the small memory Pfam readers in place and
call them "legacy". This requires keeping the old ESL_MSAFILE
structure around. I'll rename it at some point. esl-alistat opens a
new-style ESLX_MSAFILE when not using --small; with --small, it opens
a legacy ESL_MSAFILE.

If I get to the point of implementing compressed MSAs and a new
compressed MSA file format, the need for the --small Pfam readers
might go away.

Added a workaround for an Rfam 10 issue in the Stockholm parser.  Rfam
is using lines of the format "#=GF NC undefined". Parser is expecting
NC lines to have one (Rfam) or two (Pfam) real-valued
thresholds. Added a check for "undefined" string as well, in which
case NC threshold is left unset. This only affects NC lines, not TC or
GA.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@712 db3e70e4-12e1-0310-b23f-81a6e3d09ba7