bugfix in ESL_BUFFER.
It was possible to put two anchors at the same buffer position, for
example in esl_msafile_stockholm_GuessAlphabet() followed by the first
esl_msafile_GetLine(). Then RaiseAnchor() in GetLine() raises both
anchors, and the buffer may be advanced by the GuessAlphabet() routine
despite its request for the anchor. This sort of thing causes failures
in alphabet and format guessing in nonrewindable streams such as stdin
and .gz.

Fix is to add an <nanchor> field, the number of anchors at the same
position, which works like a reference counter in garbage
collectors. Setting an anchor increments it; raising one decrements
it.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@716 db3e70e4-12e1-0310-b23f-81a6e3d09ba7