Fixed problem with alphabet handling in esl-alimask miniapp. Since the msafile overhaul in revision 715, esl-alimask has been broken when used without specifying an alphabet type (--amino, --rna, or --dna) because MSAs are read in text mode and no alphabet was being defined, yet one is needed.
When not run in small-memory mode (with --small), the alphabet
is only necessary for defining gaps and so any alphabet type is
sufficient. So, we assert eslRNA unless --small is set, in which case
--amino, --dna or --rna is required.

The perl test script esl-alimask.itest.pl was not catching this
because it uses tiny alignments and so always asserts --rna. This may
no longer be necessary with the new msafile code. I'll check with SRE.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@733 db3e70e4-12e1-0310-b23f-81a6e3d09ba7