Bug #e7 fix: with icc v12, esl_random() sequence is changed.
When we upgraded to Intel icc version 12, in HMMER, the nhmmer_generic
unit test failed. Travis traced it to esl-shuffle generating a
different sequence, despite fixed RNG seed. The problem only occurs
with icc -O -fPIC.

In esl_random.c::mersenne_fill_table(), the MT incantation is
   y = (r->mt[z] & 0x80000000) | (r->mt[z+1] & 0x7fffffff);
   r->mt[z] = r->mt[z-227] ^ (y>>1) ^ mag01[y & 0x1];
with y declared as an unsigned 32-bit int. 

The problem is mag01[y & 0x1]. I think icc12 *may* (depending on
optimization and other flags) be casting y to a signed int before it
does the binary AND. The result of casting a large unsigned y to a
signed int is implementation-defined and unsafe, so if the compiler
does indeed try to cast y to signed int we're in trouble.

The fix is to dictate the cast explicitly:
   r->mt[z] = r->mt[z-227] ^ (y>>1) ^ mag01[(int)(y & 0x1)];
and this should be guaranteed to work, because we know that the result
of (y & 0x1) is 0 or 1, and C99 guarantees that we can safely cast it
to signed int 0 or 1.

Arguably this is a compiler bug. The C99 standard says that an array
index may be any integer type; it does not specify signed or unsigned,
so I don't believe the compiler is permitted to do an unsafe cast to
signed int.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@738 db3e70e4-12e1-0310-b23f-81a6e3d09ba7