Added esl_scorematrix_JointToConditionalQuery(): implements correct calculation of conditional probs P(b|a) used by HMMER single sequence queries; P(b|a) = P(ab) / P(a), where P(a) = \sum_b P(ab), and P(a) is not necessarily equal to the background frequency f_a. Using f_a (as before in HMMER) is not correct.
Fixed errors in documentation of esl_scorematrix_ProbifyGivenBG(). Was
claiming that it returned a KxK matrix over only the canonical
residues. It actually returns a Kp x Kp matrix, inclusive of
degeneracy codes (obtained by the appropriate marginalizations over
canonical residues), with the convention that all values involving gap
-, nonresidue *, and missing data ~ are 0.0. This documentation error,
combined with the lack of a routine for calculating the conditionals
P(b|a) correctly, likely contributed to a serious HMMER bug, where HMMER
was calculating P(b|a) using f_a, and only over canonical residues.




git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@749 db3e70e4-12e1-0310-b23f-81a6e3d09ba7