Improved autodetection of A2M format in eslx_msafile_GuessFileFormat(). Require that an A2M file must have nonzero number of consensus columns. Otherwise an unaligned lower-case FASTA file gets detected as an A2M alignment of all inserts and zero consensus columns.
Added utest_tricky_format_decisions() to test for this, and to be a
receptacle for future such cases.

The format guesser is also now a little more robust against unaligned
FASTA files, by adding a positive test for unaligned FASTA, for which
we return an eslENOFORMAT error. (Previously, if we failed to
positively id A2M vs. AFA, we assumed AFA -- because the routine is
built to assume that the caller has already promised us that it's an
alignment file, and the routine is only tasked with "given that this
is an alignment, what format is it?")



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@757 db3e70e4-12e1-0310-b23f-81a6e3d09ba7