Related to hmmer r4044.
Added new nhmmer-specific architecture to assign a sequence id 
to each sequence block, then use that to assign to each hit the
length of the db sequence from which it came. This is used to 
print out the db sequence length in a field "sq-len" for each
hit in dfamtblout format.  This is useful both for Robert Hubley
(RepeatMasker) processing and for identifying hits on reversed
sequence that shouldn't be treated as false positives because they
are actually self hits.  This depends on capturing lengths in 
id_length_list, then using this list to fill in db-seq-lengths
during postprocessing.

Modify ReadBlock approach, so that the sequence chunks read in 
single- and multi-threaded variants always agree. Involves 
setting and using new length limits for ReadBlock's ReadWindow
call.


esl_sqio_ascii.c : * help gather sequence lengths for the id_length pairs
                   * new length limits in readBlock()
esl_sqio_ncbi.c  : same




git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@763 db3e70e4-12e1-0310-b23f-81a6e3d09ba7