Removed the --tabfile, --shortname, --Tmin, and --Emax options from esl-sfetch, which allowed the program to fetch subsequences (hits) listed in an Infernal v1.0-->v1.0.2 cmsearch "--tabfile" output file. That cmsearch option no longer exists in Infernal v1.1, so these options wouldn't work with the new version.
I considered updating the options to work with the new --tblout
tabular format, but I now think that having these types of options in
esl-sfetch is unnecessarily complex. The only users who would notice
and use them would be users with some unix competence who I think it's
safe to assume can figure out how to manipulate the tabular --tblout
output into the GDF format that esl-sfetch -Cf accepts. And I've put
an example of how to do this in the Infernal user guide. 

Removing these options gets rid of about 200 of the 700 lines of code
from esl-sfetch.c.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@768 db3e70e4-12e1-0310-b23f-81a6e3d09ba7