Changed version from 0.1.snap20080611 to i1.1rc1, and date to June 2012.
easel's configure is not normally run in an build of Infernal, but it
can be if a sneaky user goes into infernal/easel/ and manually runs
configure. 

With the change in this revision, Easel miniapp programs that get
built following an infernal/easel/configure will have version info set
as i1.1rc1, not 0.1.snap20080611.

(Remember this is not the trunk, but an infernal release branch).


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1@773 db3e70e4-12e1-0310-b23f-81a6e3d09ba7