Merged changes from Infernal 1.1 branch back to trunk by executing the following command:
svn merge -r 773:774 https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1

Change is to check if documentation directory exists before trying to
build a Makefile in it. This directory gets removed in a distro
release of HMMER or Infernal, so it's important configure doesn't try
to create a Makefile in there (if it does, configure will fail).


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@776 db3e70e4-12e1-0310-b23f-81a6e3d09ba7