Merging r779 changes from Easel's Infernal 1.1 release branch back to trunk.
Log from r779 in Easel's Infernal 1.1 branch:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Related to r4154 changes in Infernal 1.1 release branch, and r4156 in
HMMER's Infernal 1.1 release branch, fixing a problem with 'make
clean' and 'make distclean' in a released distribution of both
Infernal (as of 1.1) and HMMER (as of next release).

Previously, a released distribution of HMMER or Infernal did not
include the easel/documentation/ directory, because building the guide
requires the executable programs 'rman', 'pdflatex', 'bibtex' to be in
the user's path.

As of now, a released distribution will include the
easel/documentation directory, but a 'make' in that directory will
first check to see if 'rman', 'pdflatex' and 'bibtex exist in the
user's path using the 'command -v' command. As far as I can tell using
'command' is the safest way to do this, although another option is to
run each command with the --version option and checking if the return
status is in error or not.

configure.ac changes because previously we had to ensure
documentation existed before trying to create it's makefile
with AC_CONFIG, now we don't have to check because it will always
exist. 
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@780 db3e70e4-12e1-0310-b23f-81a6e3d09ba7