Merged r783 changes from Infernal 1.1 release branch to Easel trunk with command:
svn merge -r 782:783 https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1

Log from r783 in Infernal 1.1 release branch of Easel:

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Fixed bug in esl-alimanip.c: --seq-k, --seq-r and --reorder only work
if msa->index (sequence name keyhash) is filled. These options remove
a subset of sequences, keep only a subset of sequences, or reorder
sequences in a MSA.

msa->index is filled for Stockholm alignments, but not AFA alignments
(and probably other formats). This meant that --seq-k, --seq-r and
--reorder would not work on AFA alignments. The 'fix' is to check that
the alignment file is Stockholm if one of these options is used and
exit with an error message if it isn't. I think this is okay because
these are 'expert' options and it's okay to require the expert user
reformat to stockholm before using them -- they can always reformat
back to the desired format afterwards.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@784 db3e70e4-12e1-0310-b23f-81a6e3d09ba7