Important bugfix in esl_sse_expf(). min/max ranges were screwed up.
It was allowing -103.279 <= x <= 88.723, corresponding to log(2^-149)
<= x <= log(2^128), which seemed to make sense because it's the widest
possible representation of e^x in single-precision floats, including
subnormals.

I failed to see that the algorithm calculates expf(x) ~ 2^k expf(f),
where the 2^k piece is constructed by bit twiddling. To make it work,
k must be in range 0..255, because we construct an IEEE754 float with
it, with sign bit 0, 8-bit exponent (k+127)<<23, and mantissa 0. This
means that the calculation gives:

  x/log2 = ...  -127.5 .. -126.5 ........ 126.5 .. 127.5 ...
        INVALID |-- e^x = 0 ---| e^x=e^x |-- e^x= +inf | INVALID 

so, if we place our min/max bounds anywhere in those 0,+inf zones, we
get 0,+inf either from our min/max bound or from letting the
calculation proceed. So we want to choose minlogf, maxlogf safely
inside the 0,inf zones, enough inside them to avoid any nasty roundoff
problems. Pommier used epsilon ~ 3e-6, and used: 
  minlogf = -127.5log2 + epsilon = -88.3763 
  maxlogf  = 127.5log2 - epsilon = 88.3763.
(i.e. Pommier had thought this through, and I didn't understand
what he was doing.)

The bug manifested itself when sampling a Dirichlet produced a
subnormal probability p, log p fell inside the -103.279..-88.3763
range between bad and correct minlogf, and a subsequent expf(log p)
had the buggy calculation trying to run, corrupting the 2^k
exponent. That is: any exp(log p) for p ~ FLT_MIN (1e-38) or less was
being corrupted. Fortunately it's unlikely that any HMMER3/Infernal
probability is ever that small, except in a unit test (like the one I
was writing today) that really stresses the extremes, with subnormal
probabilities, which is probably why we haven't seen this bug before
now; it's been in the code since Dec 2007. (And it's essentially
impossible that we'd ever see an odds ratio around FLT_MAX, so we
never saw bugs in the very narrow positive problematic region
x=88.376..88.723 either.)


Also: changed the behavior of --enable-debugging. Now it defines
eslDEBUGLEVEL to 1, not 0; so the minimum debugging level is 1.  This
makes ESL_DASSERT1() macros run at minimal debugging level.
ESL_DASSERT1() becomes a good way to insert argument "contract
checks", that will be conditionally compiled and used only when
--enable-debugging is on.





git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@793 db3e70e4-12e1-0310-b23f-81a6e3d09ba7