Merged changes from Infernal 1.1 release branch of Easel with command:
svn merge -r 793:794
https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1

Log from r794 in Infernal 1.1 release branch of Easel:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Added support for zero-length sequences in
sqascii_ReadBlock(). Previously these weren't expected or handled
gracefully, but now the function assumes they may exist and checks for
them explicitly. Now, some sequences in a ESL_SQ_BLOCK returned from
this function may have a length of 0. This commit is tied to r4196 of
the Infernal 1.1 branch of infernal as it is necessary to fix bug i30
in infernal - failure of cmsearch to be able to cope with 0 length
sequences.

This commit is in the Infernal 1.1 release branch of Easel. I'll merge
it to the trunk shortly.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@795 db3e70e4-12e1-0310-b23f-81a6e3d09ba7