Fixed a bug reported by Rob: esl-reformat --mingap corrupts SS/SS_cons annotation for protein alignments.
Problem arises from an earlier bugfix. Functions that remove gap
columns can break RNA SS annotation by removing only one side of an
annotated bp. (This includes esl_msa_ColumnSubset() and anything that
calls it, like esl_msa_MinimGaps(), esl_msa_NoGaps().) To fix this, we
made esl_msa_ColumnSubset() call esl_msa_RemoveBrokenBasepairs(). But
that assumes that all MSAs use WUSS notation in their SS/SS_cons
lines, and only nucleic acid alis use WUSS. Protein SS annotation gets
corrupted, converted to WUSS.

Fixing it is complicated by the fact that we call these functions on
text mode alignments, and our intent with text mode is that we assume
a text-mode alignment uses an arbitrary (and unknown) alphabet -- by
default, we're not allowed to make assumptions about any symbol in a
text mode alignment. So we should have to provide an affirmative flag
(do_fixbps, for example) to trigger editing of an SS annotation in a
text mode alignment when we're somehow told that it's RNA/DNA. But I'd
rather not add another arg to the interfaces for these esl_msa_*()
functions.

So the fix is as follows:

- Add esl_msa_MinimGapsText(), esl_msa_NoGapsText(). These are specialized
  to handling the text mode ali case, and they take a <do_fixbps?> flag.

- esl-reformat miniapp, which works in text mode by design, calls
  these text mode functions.  If reformat has affirmative evidence
  that it's reformatting a nucleic acid alignment with WUSS notation
  -- if options have set one or more of <force_rna>, <force_dna>,
  <wussify>, <dewuss>, <fullwuss> -- then it sets <do_fixbps> TRUE.

- esl_msa_MinimGaps() and esl_msa_NoGaps() now call the text functions,
  for text mode msa's, with do_fixbps = FALSE. This changes their 
  previous behavior on text mode msa's, prohibiting them from 
  altering SS annotations on an arbitrary-alphabet msa. 

This fixes Rob's bug, and breaks esl-alimask and its unit test, which
assume that any MSA will have its SS annotation in WUSS. So also:

- esl_msafile2_RegurgitatePfam() changed so that it only calls
  RemoveBrokenBasepairs if it affirmatively knows the alignment is
  DNA or RNA. (This is used in alimask's --small mode.)

- esl_alimask changed so when it's not in --small mode, it
  calls RemoveBrokenBasepairs() before ColumnSubset(), if a
  nucleic acid alphabet has been set with --rna or --dna options.

- esl_alimask_itest.pl driver changed so when it tests --amino,
  it no longer assumes that the SS is changed to wuss format (!);
  SS annotation symbols may not change in the --amino case.





git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@809 db3e70e4-12e1-0310-b23f-81a6e3d09ba7