Updated esl-alimerge so it can handle local alignments from Infernal (cmalign or cmsearch -A). Previously, it would fail on these alignments because it didn't know how to handle the RF '~' columns that Infernal uses to annotate columns containing residues emitted from an EL (local end) state.
This revision is related to Infernal 1.1 release branch revision 4271
which enforced a convention in Infernal for always placing EL columns
5' of insert columns. This convention simplifies the code for
merging such alignments in esl-alimerge. 

Note that currently esl-alimerge does *not* take special care to be
consistent with the HMMER3 convention of using '~' to annotate
fragment ends in *sequences* in hmmbuild input/output alignments. This
means that if esl-alimerge is used to merge alignments with '~'
fragment annotation, it will 'break' that annotation by interspersing
gap characters. For example:

Merging: 

seq1     ~~~AAA~~~
#=GC RF  ...xxx...

with 

seq2     ~~~~CCC~~
#=GC RF  ....xxx..

would result in:

seq1     .~~~AAA~~~
seq2     ~~~~CCC~~.
#=GC RF  ....xxx...

This is 'broken' because hmmbuild could not read this resulting
alignment, since it doesn't follow HMMER's convention for marking
fragments with contiguous stretches of '~' at the beginning and end of
each fragment. 

I could fix this, so resulting output alignment will be:

seq1     ~~~~AAA~~~
seq2     ~~~~CCC~~~
#=GC RF  ....xxx...

However, I think users would only rarely run up against this, and so
I'm not sure it's worth adding the extra code to deal with. I'll
discuss this with Sean.

This highlights the different use of the missing data symbol '~' in
HMMER and Infernal. In HMMER (as far as I know) it is used solely in
sequences, to mark fragments. In Infernal it is used solely in RF and
SS_cons annotation to annotate columns that include residues emitted
from an EL state. Since they're disjoint, esl-alimerge could be
consistent with both conventions.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1@811 db3e70e4-12e1-0310-b23f-81a6e3d09ba7