sqc updated. Want to be able to run MPI or valgrind tests only on test platforms where we have MPI or valgrind installed. Previously we'd been using SQCLEVEL=3 to keep valgrind tests from running on other platforms, but that's not what SQCLEVEL was intended for. sqc -M, -V options tell sqc that MPI, valgrind tests can be run; sqc then verifies that this is true, by checking that `valgrind --version` or `mpiexec -V` will run. (Note that `mpiexec` is specified by the MPI spec, but its options are not. `mpiexec -n <numprocs> <program>` is the spec's "suggestion". `mpiexec -V` happens to work to get version info on both OpenMPI and IntelMPI, but I haven't checked how portable that actually is. Doesn't matter too much -- we're not (yet) intending for MPI tests to run as part of the distro `make check`.)
Setting platform-specific options for `sqc` is inconvenient when `sqc`
is in a Makefile, as in `make check`, so a second mechanism is
provided: setting environment variables SQC_VALGRIND, SQC_MPI has the
effect of -V, -M. 

An "mpi" exercise type is now added. This is the same as an sqc
"exercise" except that such tests are skipped unless -M or SQC_MPI is
set.

Handling of "valgrind" test types is modified, so these tests are
skipped unless -V or SQC_VALGRIND is set.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@831 db3e70e4-12e1-0310-b23f-81a6e3d09ba7