Minor bug fix for input of '0' as <start> or <end> coord with -c/-C <start>..<end> when fetching subsequences. If '0' is entered as either <start> or <end>, either on command line or in a file with -f option, the program now exits with an error message specifying only positive integers are allowed.
Previously, a '0' as start coord would (at least occasionally) cause a
seg fault, and a '0' as end coord would be interpreted as the length
of the sequence (fetching to the end), which is undesirable,
e.g. '100..0', would fetch from '100..L' rather than '100..1'.

Fix is in esl-sfetch.c::parse_coord_string() and
esl-sfetch.c::multifetch_subseq() which now check for '0' case
and call esl_fatal if it's found. 

This change won't affect any other programs besides esl-sfetch that
use Easel's sequence fetching code. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@845 db3e70e4-12e1-0310-b23f-81a6e3d09ba7